requirements.txt
boto3==1.42.23
botocore==1.42.23
Flask==3.1.2
flask_cors==6.0.2
numpy==2.4.2
Pillow==12.1.0
protobuf==6.33.5
PyJWT==2.10.1
python-dotenv==1.2.1
python_bcrypt==0.3.2
Requests==2.32.5
torch==2.7.1+cu118
transformers==4.57.3
